Refactor of the PopUpCard component
This refactor shows how a component should only need the exact data
required to render it's visuals. It doesn't care about HOW it gets the
data, or where it comes from. It just renders what it receives.

The parent (in this case Dashboard) is who sends the data down to the
component.

Since Dashboard has already been hooked up to the state and called the
dispatch, it's job is to pass down the shipment to the card.

Eventually we will have some sort of MapView which we pass the
shipments to, and when you click on a marker on the map...it will have
some logic to pull the correct shipment from it's array passed down to
it down into the popup.

Hope that makes sense. I can go over this code later if you want some
more explanation.